backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.18
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
slowapi>=0.1.9
boto3>=1.34.0
jinja2>=3.1.4
email-validator>=2.1.1
python-dotenv>=1.0.1
